pyproject.toml
[tool.poetry]
name = "user-microservice-python"
version = "0.1.0"
description = "This is a microservice that represents the users of the application"
authors = ["gustavovalle23 <gustavo.vallerp@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.2"
Django = "^4.1"
django-extensions = "^3.2.0"
ipython = "^8.4.0"
pytest-django = "^4.5.2"
djangorestframework = "^3.13.1"
Markdown = "^3.4.1"
django-filter = "^22.1"
djangorestframework-simplejwt = "^5.2.0"
pre-commit = "^2.20.0"
freezegun = "^1.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
